Add 404 page for invalid spring mappings
Updated our configuration so requests for unknown url mappings are
directed to a 404.jsp page with the correct http headers set.

This required updates to our view resolver to serve JSP views and
updates to the react page controller to directly forward to index.html
instead of using the view resolver.

Also created a 404.jsp file styled to match our OL theme.

refs #12480